% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{batch.analysis.ISCCD}
\alias{batch.analysis.ISCCD}
\title{Analyze the inter-sample correlation and distance}
\usage{
batch.analysis.ISCCD(ES, s1.cols, s2.cols, s1.label = "s1", s2.label = "s2",
  file.prefix = NULL, fig.res = 150)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{s1.cols}{(numeric) the column indices for the first set of cells}

\item{s2.cols}{(numeric) the column indices for the second set of cells}

\item{s1.label}{(character) the name for the first set of cells}

\item{s2.label}{(character) the name for the second set of cells}

\item{fig.prefix}{(character) the file prefix of the output figure; if file.prefix is null, the figure will be plotted to the screen}
}
\description{
Analyze the inter-sample correlation and distance
}
