% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{batch.analysis.MA}
\alias{batch.analysis.MA}
\title{plot the MA plot for two sets of cells}
\usage{
batch.analysis.MA(x, s1.cols, s2.cols, fig.filename = NULL,
  fig.main = "MA plot", fig.res = 150)
}
\arguments{
\item{x}{The expression matrix; rows are genes, cells are columns}

\item{s1.cols}{(numeric) the column indices for the first set of cells}

\item{s2.cols}{(numeric) the column indices for the second set of cells}

\item{fig.filename}{(character) the file name of the output figure; if fig.filename is null, the figure will be plotted to the screen}

\item{fig.main}{(character) the title of the figure}

\item{s1.label}{(character) the name for the first set of cells}

\item{s2.label}{(character) the name for the second set of cells}

\item{do.log}{(logical) log2 transformation of mean expression values of each gene}
}
\description{
plot the MA plot for two sets of cells
}
