% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{batch.analysis.QQ}
\alias{batch.analysis.QQ}
\title{plot the QQ plot for two sets of cells}
\usage{
batch.analysis.QQ(ES, s1.cols, s2.cols, do.log2 = TRUE, fig.filename = NULL,
  s1.label = "s1", s2.label = "S2", fig.main = "QQ plot", fig.res = 150)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{s1.cols}{(numeric) the column indices for the first set of cells}

\item{s2.cols}{(numeric) the column indices for the second set of cells}

\item{do.log2}{(logical) log2 transformation of mean expression values of each gene}

\item{fig.filename}{(character) the file name of the output figure; if fig.filename is null, the figure will be plotted to the screen}

\item{s1.label}{(character) the name for the first set of cells}

\item{s2.label}{(character) the name for the second set of cells}

\item{fig.main}{(character) the title of the figure}
}
\description{
plot the QQ plot for two sets of cells
}
