% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{batch.analysis}
\alias{batch.analysis}
\title{Analyzing batch difference}
\usage{
batch.analysis(object, analysis = c("q", "qq", "ma", "ngenes",
  "distribution"), min.expression = 1, ...)
}
\arguments{
\item{object}{A sincera object}

\item{analysis}{A vector containing the analyses to be performed, including
"q" - plotting the quantiles of gene expression in individual cells
"qq" - Q-Q plot per sample pair
"MA" - MA plot per sample pair
"ngene" - Number of expressed genes in cells of different batches
"distribution" - The distribution of expression values in cells of different batches}

\item{min.expression}{The threshold of expressed genes}
}
\value{
The updated sincera object with batch analyses results
}
\description{
Analyzing batch difference
}
