% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{celltype.enrichment.initKB}
\alias{celltype.enrichment.initKB}
\title{Contruct the knowledge base for cell type enrichment analysis}
\usage{
celltype.enrichment.initKB(associations, genome, species = "MUSMU",
  id.type = "ENSEMBL", verbose = T)
}
\arguments{
\item{associations}{(data.frame) a data frame containing the gene and cell type associations}

\item{genome}{(character) the full set of genes in a scRNA-seq data}

\item{species}{(character) the species of the genome: MUSMU - mouse, HOMSA - human}

\item{id.type}{(character) the type of ids of the genes in the genome: ENSEMBL - ENSEMBL gene id, SYMBOL -  Entrez Gene Symbol, EG - Entrez Gene Id}

\item{verbose}{(logical)}
}
\value{
a list of 3 items: celltype.gene.association - genome related associations, celltype.genome.count - genome-wide cell type associations, genome - symbol mapping for the genome
}
\description{
Contruct the knowledge base for cell type enrichment analysis
}
