% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{celltype.enrichment.old}
\alias{celltype.enrichment.old}
\title{Contruct the knowledge base for cell type enrichment analysis}
\usage{
celltype.enrichment.old(ES, KB = NULL, group.by = "CLUSTER",
  groups = NULL, species = "MUSMU", id.type = "SYMBOL",
  celltype.enrichment.prefix = CELLTYPE.ENRICHMENT.PREFIX, verbose = TRUE)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{KB}{(list) the knowledge base prepared by celltype.enrichment.initKB()}

\item{group.by}{(character) the name of the column that contains the cluster information}

\item{groups}{(character) the clusters for cell type enrichment}

\item{species}{(character) the species of the genome: MUSMU - mouse, HOMSA - human}

\item{id.type}{(character) the type of ids of the genes in the genome: ENSEMBL - ENSEMBL gene id, SYMBOL -  Entrez Gene Symbol, EG - Entrez Gene Id}

\item{celltype.enrichment.prefix}{(character) the prefix of columns encoding the cluster-specific gene list for cell type enrichment analysis}

\item{verbose}{(logical)}
}
\description{
Contruct the knowledge base for cell type enrichment analysis
}
