% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{celltype.validation.old}
\alias{celltype.validation.old}
\title{Cell type validation using the expression patterns of known biomarkers}
\usage{
celltype.validation.old(ES, group.by = CLUSTER.LABEL, groups = NULL,
  threshold = 0, marker.prefix = MARKER.PREFIX, verbose = TRUE)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the cluster information}

\item{groups}{(character) the clusters for cell type validation}

\item{threshold}{(numeric) the threshold of expression}

\item{marker.prefix}{(character) the prefix of columns encoding the cell type markers}

\item{verbose}{(logical)}
}
\description{
Cell type validation using the expression patterns of known biomarkers
}
