% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{cluster.assignment.old}
\alias{cluster.assignment.old}
\title{Assign cells to clusters}
\usage{
cluster.assignment.old(ES, cluster.label = CLUSTER.LABEL,
  clustering.method = "hc", distance.method = "pearson",
  linkage.method = "average", num.singleton = 0, h = NULL, k = NULL,
  do.shift = TRUE, do.plot = TRUE, target = 1, k.min = 25,
  min.area.increase = 0.2, maxK = 10, reps = 10, pItem = 0.8,
  pFeature = 1, clusterAlg = "hc", verbose = TRUE, export = TRUE,
  export.components = "pd")
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{clustering.method}{(character) the clustering method: hc - hierarchical clustering, tight - tight clustering, consensus - consensus clustering}

\item{distance.method}{(character) the distance method used in the hierarchical and consensus clustering: pearson - (1-pearons'correlation)/2, spearman - (1-spearman's correlation)/2, euclidean - euclidean distance}

\item{num.singleton}{(numeric) when hierarchical clustering is used, the number of singleton clusters that will be allowed in the generated clustering scheme}

\item{h}{(numeric) parameter for hierarchical clustering; the distance threshold that will be used to generate cell clusters}

\item{k}{(numeric) parameter for hierarchical clustering; the number of clusters that will be generated}

\item{do.shift}{(logical) parameter for hierarchical clustering}

\item{do.plot}{(logical) parameter for hierarchical clustering; plot the hierarchical tree}

\item{target}{(numeric) the targeted number of clusters for tight clustering; required for tight clustering}

\item{k.min}{(numeric) required for tight clustering}

\item{maxK}{(numeric) parameter for consensus clustering}

\item{reps}{(numeric) parameter for consensus clustering}

\item{pItem}{(numeric) parameter for consensus clustering}

\item{pFeature}{(numeric) parameter for consensus clustering}

\item{clusterAlg}{(character) the clustering method used in the consensus clustering}

\item{verbose}{(logical)}

\item{export}{(logical) export the clustering}

\item{export.components}{(character) the components of an ExpressionSet object that will be exported}

\item{cluster.lable}{(character) the name of the column that will be used to encode the result of cluster assignment}

\item{min.are.increase}{(numeric) required for consensus clustering}
}
\value{
a list of two elements: cell.cluster - a vector encoding the result of cluster assignment; cell.order - when hc is used, the cells in the order as they appeared in the hierarchical tree
}
\description{
Assign cells to clusters
}
