% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{cluster.diffgenes}
\alias{cluster.diffgenes}
\title{Detect cluster specific differentially expressed genes}
\usage{
cluster.diffgenes(object, groups = NULL, genes = NULL, method = "welch",
  do.fdr = FALSE, thresh = 0.05, ...)
}
\arguments{
\item{object}{A sincera object}

\item{groups}{The cell groups included in the differential expression analysis; if NULL, set to all groups defined in the object}

\item{genes}{The set of genes included in the analysis; if NULL, set to all genes in the object}

\item{method}{The method for differential test, possible values include welch - one tailed welch's t-test, wilcoxon - one-tailed wilcoxon rank sum test}

\item{do.fdr}{If TRUE, do the FDR correction}

\item{thresh}{The threshold for significance}
}
\value{
The update sincera object with diff test results in the difftests slot and the significant diff genes in the diffgenes slot
}
\description{
Detect cluster specific differentially expressed genes
}
