% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{cluster.geneSelection}
\alias{cluster.geneSelection}
\title{Selecting genes for cell cluster identification}
\usage{
cluster.geneSelection(object, method = "specificity", pergroup = TRUE,
  min.samples = 2, specifity.thresh = 0.7, do.plot = T, ...)
}
\arguments{
\item{object}{A sincera object}

\item{method}{The selection method, possible values include specificity}

\item{pergroup}{If TRUE, calculate metrics for each cell groups (samples)}

\item{specifity.thresh}{The specificity threshold}

\item{do.plot}{If TRUE, plot figures of selection}

\item{min.sample}{The selected genes must pass the selection criteria in at least min.sample samples}
}
\value{
The updated sincera object with the selected genes in the genes.forclustering slot
}
\description{
Currently support the specificity metric (Guo et al., PLoS Comp Bio 2015); other metrics will be supported soon
}
