% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{cluster.ordering}
\alias{cluster.ordering}
\title{Ordering rows or columns according to a specified order}
\usage{
cluster.ordering(ES, col.order = NULL, row.order = NULL, verbose = TRUE)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{col.order}{(numeric) the names of columns in the new order}

\item{row.order}{(numeric) the names of rows in the new order}

\item{verbose}{(logical)}
}
\value{
an ExpressionSet object with re-ordered data
}
\description{
Ordering rows or columns according to a specified order
}
