% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{cluster.permutation.analysis.1}
\alias{cluster.permutation.analysis.1}
\title{Permutation Analysis for determining significance of cluster assignments}
\usage{
cluster.permutation.analysis.1(ES, group.by = "GROUP", n = 20,
  distance.method = "euclidean", log.base = 2, verbose = TRUE)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the cluster information}

\item{n}{(numeric) the number of permutations}

\item{distance.method}{(character) the distance method: pearson or spearman - (1-correlation)/2; euclidean - euclidean distance}

\item{log.base}{(numeric) the base of logorithm; if log.base <=1 or log.base is NULL, no log transformations will be applied}

\item{verbose}{(logical)}
}
\description{
Permutation Analysis for determining significance of cluster assignments
}
\note{
This function is from SINCERA version a10242015. It will be upgraded soon.
}
