% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{cluster.permutation.analysis}
\alias{cluster.permutation.analysis}
\title{Validating clustering results using permutation analysis}
\usage{
cluster.permutation.analysis(object, n = 20, distance.method = "euclidean",
  ...)
}
\arguments{
\item{object}{A sincera object}

\item{n}{The number of random permutations}

\item{distance.method}{The method to calculate cell expression profile distance, possible values include pearson and euclidean}
}
\description{
Validating clustering results using permutation analysis
}
