% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{consensus_maximization}
\alias{consensus_maximization}
\title{consensus maximization}
\usage{
consensus_maximization(x, prior = NULL, alpha = 4, epslon = 0.1,
  max.iter = 100)
}
\arguments{
\item{x}{- normalized model prediction}
}
\description{
consensus maximization
}
