% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{construct}
\alias{construct}
\title{Construct a sincera object}
\usage{
construct(exprmatrix = NULL, exprfile = NULL, samplevector = NULL,
  samplefile = NULL, projname = NULL)
}
\arguments{
\item{exprmatrix}{The expression matrix; rows are genes, columns are cells}

\item{exprfile}{The path to the file containing the expression matrix; either exprmatrix or exprfile must be set; if both provided, exprmatrix will be used}

\item{samplevector}{A vector containing the batch information of cells}

\item{samplefile}{The path to the file containing the batch information of cells; if both samplevector and samplefile are set to NULL, all cells will be assigned as "sample1"}

\item{projname}{A string describing the data and analysis}
}
\value{
The created sincera object
}
\description{
Construct a sincera object from data in memory or files on disk
}
