% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{copyCellMeta}
\alias{copyCellMeta}
\title{Add or update a cell meta data using a predefined cell meta data}
\usage{
copyCellMeta(object, from, to, verbose = T)
}
\arguments{
\item{object}{A sincera object}

\item{from}{The name of the predefined cell meta data}

\item{to}{The name of the meta data to be added or updated}
}
\value{
The update sincera object
}
\description{
Add or update a cell meta data using a predefined cell meta data
}
