% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{copyGeneMeta}
\alias{copyGeneMeta}
\title{Add or update a gene meta data using a predefined gene meta data}
\usage{
copyGeneMeta(object, from, to)
}
\arguments{
\item{object}{A sincera object}

\item{from}{The name of the predefined gene meta data}

\item{to}{The name of the meta data to be added or updated}
}
\value{
The update sincera object
}
\description{
Add or update a gene meta data using a predefined gene meta data
}
