% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{diff.test}
\alias{diff.test}
\title{Detecting Cluster Specific Differentially Expressed Genes}
\usage{
\method{diff}{test}(ES, group.by = CLUSTER.LABEL, groups = NULL,
  method = "welch", do.fdr = FALSE, samseq.fdr = 0.2,
  samseq.nperms = 10, samseq.nresamp = 20,
  diffexpr.prefix = DIFF.EXPR.PREFIX, verbose = T, export = T,
  export.components = "fd")
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the cluster information}

\item{groups}{(character) the clusters for detecting cluster specific differentially expressed genes}

\item{method}{(character) differential test method: welch - Welch's t-test, wilcoxon - Wilcoxon rank sum test, samseq - SAMseq}

\item{do.fdr}{(logical) performing B&H fdr when method is welch or wilcoxon}

\item{samseq.fdr}{(numeric) parameter for SAMseq}

\item{samseq.nperms}{(numeric) parameter for SAMseq}

\item{samseq.nresamp}{(numeric) parameter for SAMseq}

\item{diffexpr.prefix}{(character) the prefix to label the columns encoding the results of differential test}

\item{verbose}{(logical)}

\item{export}{(logical) wheterh to export the ExpressionSet}

\item{export.components}{(character) the components of an ExpressionSet object that will be exported}
}
\value{
an ExpressionSet object containing the results of differential tests in the attributes of fData
}
\description{
Detecting Cluster Specific Differentially Expressed Genes
}
