% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{doPCA}
\alias{doPCA}
\title{Perform PCA based dimension reduction}
\usage{
doPCA(object, genes = NULL, use.scaled = T, use.fast = T, ...)
}
\arguments{
\item{object}{A sincera object}

\item{genes}{The set of genes used for PCA}

\item{use.scaled}{If TRUE, use the scaled expression for PCA}

\item{use.fast}{If TRUE, use gmodels::fast.prcomp; otherwise, use stats::prcomp}
}
\value{
The update sincera object with PCA results in the pca slot
}
\description{
Perform PCA based dimension reduction
}
