% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{doTSNE}
\alias{doTSNE}
\title{Perform tSNE based dimension reduction}
\usage{
doTSNE(object, n = 2, use.fast = T, max.iter = 2000, seed = 0,
  feature.type = "pca", dims = 1:3, genes = NULL, use.scaled = F, ...)
}
\arguments{
\item{object}{A sincera object}

\item{n}{The number of tSNE dimensions to be extracted}

\item{use.fast}{If TRUE, use Rtsne::Rtsne; otherwise, use tsne::tsne}

\item{max.iter}{The maximum number of iteration}

\item{seed}{The seed of randomness}

\item{feature.type}{The input feature type for tSNE reduction, possible values include gene and pca}

\item{dims}{If feature.type is pca, the pca components for tSNE}

\item{genes}{If feature.type is genes, the set of genes for tSNE; if NULL, set to all genes}

\item{use.scaled}{If TRUE, use the scaled expression for tSNE}

\item{use.scaled}{If TRUE, use the scaled expression values of genes for tSNE; only take effect when the feature.type is gene}
}
\value{
The update sincera object with tsne results in the tsne slot
}
\description{
Perform tSNE analysis on gene expression space or pca space
}
