% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{driving.force.analysis.1}
\alias{driving.force.analysis.1}
\title{construct per-group transcriptional regulatory network (implementation 1)}
\usage{
driving.force.analysis.1(ES, group.by = "CLUSTER", groups = NULL,
  prefix.tg = "use_for_tg_", prefix.tf = "use_for_tf_", verbose = T)
}
\arguments{
\item{prefix.tg}{prefix for labeling the columns that specify candidate group specific regulatory targets}

\item{prefix.tf}{prefix for labeling the columns that specify candidate group specific transcriptional factors}
}
\description{
construct per-group transcriptional regulatory network (implementation 1)
}
