% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{driving.force.analysis.2}
\alias{driving.force.analysis.2}
\title{construct per-group transcriptional regulatory network (implementation 2)}
\usage{
driving.force.analysis.2(ES, group.by = "CLUSTER", groups = NULL,
  unique.by = GENE.SYMBOL.LABEL, norm.method = "zscore",
  prefix.tg = "use_for_tg_", prefix.tf = "use_for_tf_", verbose = T)
}
\arguments{
\item{prefix.tg}{prefix for labeling the columns that specify candidate group specific regulatory targets}

\item{prefix.tf}{prefix for labeling the columns that specify candidate group specific transcriptional factors}
}
\description{
construct per-group transcriptional regulatory network (implementation 2)
}
