% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{driving.force.analysis}
\alias{driving.force.analysis}
\title{per group key regulator prediction}
\usage{
driving.force.analysis(ES, group.by = "CLUSTER", groups = NULL,
  unique.by = NULL, norm.method = "zscore", prefix.tg = TG.PREFIX,
  prefix.tf = TF.PREFIX, verbose = T)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the scRNA-seq data}

\item{group.by}{(character)}

\item{groups}{(character) groups for identifying key regulators}

\item{unique.by}{(character)}

\item{norm.method}{(character)}

\item{prefix.tg}{(character) prefix for labeling columns encoding the potential regulatory targets for each group}

\item{prefix.tf}{(character) prefix for labeling columns encoding the potential transcription factors for each group}

\item{verbose}{(logical)}
}
\description{
per group key regulator prediction
}
