% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivingforce.R
\name{drivingforce.inferTRN}
\alias{drivingforce.inferTRN}
\title{Inferring significant TF->TF and TF->TG interactions based on gene expression patterns}
\usage{
drivingforce.inferTRN(object, groups = NULL, use.scaled = T,
  use.allcells = T, ...)
}
\arguments{
\item{object}{A sincera object}

\item{groups}{The cell groups for network inference}

\item{use.scaled}{If TRUE, use the zscore scaled expression values for the inference}

\item{use.allcells}{If TRUE, the expression values of all cells will be used for inference; otherwise, only the expression values in the cell group will be used for inference.}
}
\value{
An update sincera object, use getDF with name="edges" to access the significance of all interactions
}
\description{
Inferring significant TF->TF and TF->TG interactions based on gene expression patterns
}
