% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivingforce.R
\name{drivingforce.selectTGs}
\alias{drivingforce.selectTGs}
\title{Select candidate TGs for network inference and TF ranking}
\usage{
drivingforce.selectTGs(object, groups = NULL, tglists = NULL,
  diff.method = "welch", diff.thresh = 0.01, ...)
}
\description{
Please use the \"tglists\" parameter to provide pre-selected tgs
Otherwise, the function will select candidate TGs for each group using the following criteria:
* Genes differentially expressed in the cell group (pvalue < \"diff.thresh\")
}
\details{
@param object A sincera object
 @param clusters A vector of cell groups; if NULL, use all defined cell groups
 @param tglists A list containing user-provided candidate TGs for each group specified in \"groups\"
 @param diff.method The method used to perform the differential expression analysis
 @param diff.thresh The pvalue for determining differentially expressed genes
 @return An updated sincera object, use getDF with name="tgs" to access the selected TGs
}
