% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{expr.minimum}
\alias{expr.minimum}
\title{Set the lower bound of gene expression}
\usage{
expr.minimum(object, value = 0.01, ...)
}
\arguments{
\item{object}{A sincera object}

\item{value}{The lower bound value}
}
\value{
An updated sincera object with the manipulated expression values, use getExpression function to access the expression values
}
\description{
Set the lower bound of gene expression
}
