% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{expr.quantiles}
\alias{expr.quantiles}
\title{Calculating the expression quantiles per cell}
\usage{
expr.quantiles(ES, probs = c(0.25, 0.5, 0.75, 0.8, 0.85, 0.9, 0.92, 0.94,
  0.96, 0.98, 0.99, 1), lower.bound = -1, upper.bound = Inf,
  do.plot = FALSE, color.by = "SAMPLE", fig.filename = NULL,
  fig.res = 300)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{probs}{(numeric) a set of quantiles to be calculated}

\item{lower.bound}{(numeric) expression values greater than the lower.bound in each cell will be used for the per cell quantile calculation}

\item{do.plot}{(logical) plot the per cell quantiles}

\item{fig.filename}{(character) file name of the output figure; if fig.filename is null, the figure will be plotted to the screen}

\item{lower.bound}{(numeric) expression values less than the upper.bound in each cell will be used for the per cell quantile calculation}
}
\value{
an ExpressionSet object with calculated quantiles encoded in fData attributes
}
\description{
Calculating the expression quantiles per cell
}
