% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{expression.export4viz}
\alias{expression.export4viz}
\title{Export the clustering resuls for visualization using external tools, e.g., GENE-E}
\usage{
expression.export4viz(ES, signature = NULL, cell.attrs = NULL,
  file.prefix = "sincera.cluster.export", verbose = T)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{signature}{(data.frame) a data.frame object containing cell type signature, shall contain at least two columns SYMBOL and TYPE}

\item{verbose}{(logical)}

\item{cell.order}{(numeric) the names of columns in the new order}
}
\value{
an ExpressionSet object with re-ordered data
}
\description{
Export the clustering resuls for visualization using external tools, e.g., GENE-E
}
