% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{exprs.abundancy}
\alias{exprs.abundancy}
\title{Per group expression abundancy calculation}
\usage{
exprs.abundancy(ES, group.by = SAMPLE.LABEL, groups = NULL,
  abundancy.threshold = 5, abundancy.prefix = EXPR.ABUNDANCY.PREFIX)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the sample information}

\item{groups}{(vector) samples for measuring abundancy}

\item{abundancy.threshold}{(numeric) the min value of expression}

\item{abundancy.prefix}{(character) the prefix for labeling the columns encoding the results of expression abundancy calculation for each group}
}
\value{
an ExpressionSet object with calculated abundancies encoded in fData attributes
}
\description{
Per group expression abundancy calculation
}
