% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{exprs.common}
\alias{exprs.common}
\title{Identifying group specific common genes}
\usage{
exprs.common(ES, group.by = CLUSTER.LABEL, groups = NULL,
  common.threshold = 5, common.percentage = 0.8,
  common.prefix = COMMON.PREFIX)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the cluster information}

\item{groups}{(character) the clusters for signature prediction}

\item{common.threshold}{(numeric) in common gene metric, expression > COMMON.TRESHOLD will considered as expressed}

\item{common.percentage}{(numeric) in common gene metric, genes that express in >COMMON.PRECENTAGE cluster cells will be considered as a common}

\item{common.prefix}{(character) prefix for labeling the columns encoding the results of common metric for each group}
}
\value{
an ExpressionSet object containing the results of common gene metrics in the attributes of fData
}
\description{
Identifying group specific common genes
}
