% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{exprs.specificity}
\alias{exprs.specificity}
\title{Per group expression specificity calculation}
\usage{
exprs.specificity(ES, group.by = SAMPLE.LABEL, groups = NULL,
  specificity.prefix = EXPR.SPECIFICITY.PREFIX)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the sample information}

\item{groups}{(character) samples for measuring specificity}

\item{specificity.prefix}{(character) the prefix for labeling the columns encoding the results of expression specificity calculation for each group}
}
\value{
an ExpressionSet object with calculated specificities encoded in fData attributes
}
\description{
Per group expression specificity calculation
}
\details{
if a gene has zero expression across all cells, its specificity is zero
}
