% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{exprs.synthetic.similarity}
\alias{exprs.synthetic.similarity}
\title{Calculating the similarity between each gene profile and group specific synthetic reference expression profile
synthetic profile is created by setting fpkm(cluster cells)=1, fpkm(non cluster cells)=0}
\usage{
exprs.synthetic.similarity(ES, group.by = CLUSTER.LABEL, groups = NULL,
  syn.sim.prefix = SYN.SIM.PREFIX)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the cluster information}

\item{groups}{(character) the clusters for signature prediction}

\item{syn.sim.prefix}{(character) prefix for labeling the columns encoding the results of synthetic profile similarity metric for each group}
}
\value{
an ExpressionSet object containing the results of synthetic profile similarity metrics in the attributes of fData
}
\description{
Calculating the similarity between each gene profile and group specific synthetic reference expression profile
synthetic profile is created by setting fpkm(cluster cells)=1, fpkm(non cluster cells)=0
}
