% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{exprs.test.statistic.metric}
\alias{exprs.test.statistic.metric}
\title{Normalized and smoothened results of group specific differential expression test}
\usage{
exprs.test.statistic.metric(ES, group.by = CLUSTER.LABEL, groups = NULL,
  log.base = 2, test.statistic.metric.prefix = TEST.STATS.METRIC.PREFIX,
  diff.expr.prefix = DIFF.EXPR.PREFIX)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the cluster information}

\item{groups}{(character) the clusters for signature prediction}

\item{log.base}{(numeric) the base of logarithm}

\item{test.statistic.metric.prefix}{(character) prefix for labeling the columns encoding the results of test statistic metric for each group}

\item{diff.expr.prefix}{(character) prefix for labeling the columns encoding the results of differential test}
}
\value{
an ExpressionSet object containing the results of test statistic metrics in the attributes of fData
}
\description{
Normalized and smoothened results of group specific differential expression test
}
