% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{exprs.truncate}
\alias{exprs.truncate}
\title{Truncate expression values that are out of bounds}
\usage{
exprs.truncate(ES, upper.bound = NULL, lower.bound = NULL, verbose = TRUE)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{upper.bound}{(numeric) the upper bound of expression}

\item{lower.bound}{(numeric) the lower bound of expression}

\item{verbose}{(logical) verbose the function output}
}
\value{
an ExpressionSet object containing the truncated expression values
}
\description{
Truncate expression values that are out of bounds
}
