% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{exprs.unique}
\alias{exprs.unique}
\title{Identifying group specific unique genes}
\usage{
exprs.unique(ES, group.by = CLUSTER.LABEL, groups = NULL,
  unique.ratio = 2, unique.quantile = 0.85, unique.prefix = UNIQUE.PREFIX)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the cluster information}

\item{groups}{(character) the clusters for signature prediction}

\item{unique.ratio}{(numeric) parameter for unique gene metric}

\item{unique.quantile}{(numeric) parameters for unique gene metric}

\item{unique.prefix}{(character) prefix for labeling the columns encoding the results of unique metric for each group}
}
\value{
an ExpressionSet object containing the results of unique gene metric in the attributes of fData
}
\description{
Identifying group specific unique genes
}
