% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{filterContaminatedCells}
\alias{filterContaminatedCells}
\title{Detecting and removing potentially contaminated cells}
\usage{
filterContaminatedCells(object, do.rm = T, markers.1, markers.2,
  min.expression.1 = 1, min.quantile.1 = 0.99, min.expression.2 = 1,
  min.quantile.2 = 0.99, ...)
}
\arguments{
\item{object}{A sincera object}

\item{do.rm}{If TRUE, remove detected contaminants}

\item{markers.1}{The vector encoding the markers for one cell type}

\item{markers.2}{The vector encoding the markers for the other cell type}

\item{min.expression.1}{The threshold value of the expression of markers in markers.1}

\item{min.quantile.1}{The threshold value of the expression quantile of markers in markers.1}

\item{min.expression.2}{The threshold value of the expression of markers in markers.2}

\item{min.quantile.2}{The threshold value of the expression quantile of markers in markers.2}
}
\value{
An updated sincera object with detected contaminated cells removed
}
\description{
Detecting contaminated cells based on the co-expression of highly selectively markers of two different cell types, such as epithelial and endothelial
}
