% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{filterLowQualityCells}
\alias{filterLowQualityCells}
\title{pre-filtering low quality cells}
\usage{
filterLowQualityCells(object, min.expression = 1, min.genes = 500,
  do.plot = FALSE, ...)
}
\arguments{
\item{object}{A sincera object}

\item{min.expression}{The threshold value of expressed genes}

\item{min.genes}{Cells with less than this number of expressed genes will be excluded from downstream analysis}

\item{do.plot}{If TRUE, plot the distribution of library complexity}
}
\value{
An update sincera object with low quality cells pre-filtered
}
\description{
Pre-filtering low quality cells based on library complexity (i.e., the number of genes expressed in a library)
}
