% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{GeneStats}
\alias{GeneStats}
\title{Generate gene level global or per group statistics}
\usage{
GeneStats(dp, ident, groups = NULL, genes = NULL, min.expression = 1,
  min.gsize = 2, min.avg = 1, stats = c("var", "specificity", "min",
  "max", "avg", "peak1", "peak2", "welch", "ident.fc", "ident.avg", "ident.min",
  "ident.max", "ident.cnt", "ident.pct"))
}
\arguments{
\item{dp}{The expression matrix: rows are genes, columns are cells}

\item{ident}{The vector of cell identity}

\item{genes}{The set of genes to be included in the calculation; if NULL, will be automatically set to all genes in the expression matrix}

\item{min.gsize}{The minimum size of a group}

\item{min.avg}{The minimum of average expression values}

\item{stats}{A vector containing the statistics to be calculated for each gene, including}

\item{group}{The identity groups to be included in the calculation}

\item{min.exp}{The threshold vaule of gene expression}
}
\value{
A data.frame containing the results
}
\description{
Generate gene level global or per group statistics
}
