% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{get.diff.genes}
\alias{get.diff.genes}
\title{Retrieve Differentially Expressed Genes}
\usage{
get.diff.genes(ES, group.by = CLUSTER.LABEL, groups = NULL,
  diffexpr.prefix = DIFF.EXPR.PREFIX, threshold = 0.05)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the cluster information}

\item{groups}{(character) the clusters for selecting cluster specific differentially expressed genes}

\item{diffexpr.prefix}{(character) the prefix to label the columns encoding the results of differential test}

\item{threshold}{(numeric) selecting diff genes <threshold}
}
\value{
a data frame containing the retreived differentially expressed genes
}
\description{
Retrieve Differentially Expressed Genes
}
