% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{getCellMeta}
\alias{getCellMeta}
\title{Get the meta data of cells in a sincera object}
\usage{
getCellMeta(object, name)
}
\arguments{
\item{object}{A sincera object}

\item{name}{The name of the meta data}
}
\value{
A vector containing the value of the meta data of all cells
}
\description{
Get the meta data of cells in a sincera object
}
