% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{getCellNum}
\alias{getCellNum}
\title{Get the number of cells in a sincera object}
\usage{
getCellNum(object)
}
\arguments{
\item{object}{A sincera object}
}
\value{
A numeric value representing the number of cells in the sincera object
}
\description{
Get the number of cells in a sincera object
}
