% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{getCells}
\alias{getCells}
\title{get the names of cells in specified cell groups in a sincera object}
\usage{
getCells(object, groups = NULL)
}
\arguments{
\item{object}{A sincera object}

\item{groups}{The cell groups; if NULL, set to all cell groups defined in the sincera object}
}
\value{
A character vector containing cell names
}
\description{
get the names of cells in specified cell groups in a sincera object
}
