% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{getCellType}
\alias{getCellType}
\title{Get the cell types of given cell groups}
\usage{
getCellType(object, groups = NULL)
}
\arguments{
\item{object}{A sincera object}

\item{groups}{The set of cell groups whose cell types will be returned; if NULL, set to all cell groups defined in the "GROUP" metadata}
}
\value{
A character vector of cell types
}
\description{
Get the cell types of given cell groups
}
