% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{getCellTypeMarkers}
\alias{getCellTypeMarkers}
\title{Get cell type marker information defined in a sincera object}
\usage{
getCellTypeMarkers(object)
}
\arguments{
\item{object}{A sincera object}
}
\value{
A data frame containing the markers and their corresponding cell types. The first column is cell type, and the second column is marker gene name.
}
\description{
Get cell type marker information defined in a sincera object
}
