% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{getExpression}
\alias{getExpression}
\title{get expression matrix from a sincera object}
\usage{
getExpression(object, scaled = FALSE, genes = NULL, cells = NULL)
}
\arguments{
\item{object}{A sincera object}

\item{scaled}{If TRUE, return the zscore scaled expression values}

\item{genes}{The set of genes to be returned; if NULL, set to all genes in the object}

\item{cells}{The set of cells to be returned; if NULL, set to all cells in the object}
}
\value{
A matrix containing the expression values; rows are genes, columns are cells.
}
\description{
get expression values of a set of genes in a set of cells from a sincera object
}
