% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{getExprQuantiles}
\alias{getExprQuantiles}
\title{Get and plot the quantiles of gene expression per cell}
\usage{
getExprQuantiles(object, probs = c(0.25, 0.5, 0.75, 0.8, 0.85, 0.9, 0.92,
  0.94, 0.96, 0.98, 0.99, 1), lower.bound = -1, upper.bound = Inf,
  do.plot = T)
}
\arguments{
\item{object}{A sincera object}

\item{probs}{a vector of quantiles to be computed}

\item{lower.bound}{The minimum value for quantile calculation}

\item{upper.bound}{The maximum value for quantile calculation}

\item{do.plot}{If true, visualize the quantiles in individual cells with colors indicating different batches}
}
\value{
A data frame containing the quantile values of individual cells
}
\description{
Enable analysis of batch difference by calculation and plotting the quantiles of cellular expression profiles
Colors indicate cells from different batches
}
