% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{getGeneMeta}
\alias{getGeneMeta}
\title{Get a gene meta data}
\usage{
getGeneMeta(object, name)
}
\arguments{
\item{object}{A sincera object}

\item{name}{The name of the meta data}
}
\value{
A vector encoding the value of the meta data of genes
}
\description{
Get a gene meta data
}
