% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{getGeneNum}
\alias{getGeneNum}
\title{Get the number of the genes in a sincera object}
\usage{
getGeneNum(object)
}
\arguments{
\item{object}{A sincera object}
}
\value{
The number of genes
}
\description{
Get the number of the genes in a sincera object
}
