% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{getGenes}
\alias{getGenes}
\title{Get the names of all genes in a sincera object}
\usage{
getGenes(object)
}
\arguments{
\item{object}{A sincera object}
}
\value{
A character vector of the names of all genes in the sincera object
}
\description{
Get the names of all genes in a sincera object
}
