% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{getGenesForClustering}
\alias{getGenesForClustering}
\title{Get the set of genes used for cell cluster identification}
\usage{
getGenesForClustering(object)
}
\arguments{
\item{object}{A sincera object}
}
\description{
Return the value of the genes.forclustering slot
}
