% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{getGraph}
\alias{getGraph}
\title{load graph from edgelist file and extract the largest connected component}
\usage{
getGraph(filename, threshold, directed = FALSE)
}
\description{
load graph from edgelist file and extract the largest connected component
}
