% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{getPCA}
\alias{getPCA}
\title{Get the PCA results stored in a sincera object}
\usage{
getPCA(object, name)
}
\arguments{
\item{object}{A sicnera object}

\item{name}{The name of the item to be returned, possible values include rds, loadings, sdev}
}
\value{
The value of the specified PCA results
}
\description{
Get the PCA results stored in a sincera object
}
