% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivingforce.R
\name{network.inference}
\alias{network.inference}
\title{Inferring first order conditional dependencies between TFs and TGs
Adapted from G1DBN package}
\usage{
network.inference(data, method = "ls", predPosition = NULL,
  targetPosition = NULL, lag = 0)
}
\arguments{
\item{data}{(matrix) the expression data}

\item{method}{(character) regression method: ls, turkey, huber}

\item{targetPosition}{(numeric) indices of regulatory targets}

\item{lag}{(numeric) time lag; 0 for non-time series data}

\item{prePosition}{(numeric) indices of regulators}
}
\value{
a list of score matrices
}
\description{
Inferring first order conditional dependencies between TFs and TGs
Adapted from G1DBN package
}
