% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{normalization.log.old}
\alias{normalization.log.old}
\title{log transformation}
\usage{
normalization.log.old(ES, log.base = 2, min.impute = 0.01, increase = 1)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{log.base}{(numeric) the base of the logarithm}

\item{min.impute}{(numeric) a numeric value to impute the min expression values}
}
\value{
an ExpressionSet object with log transformed expression values
}
\description{
log transformation
}
