% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{normalization.log}
\alias{normalization.log}
\title{log transformation}
\usage{
normalization.log(x, log.base = 2, min.impute = 0.01, increase = 1)
}
\arguments{
\item{x}{The expression matrix; rows are genes, cells are columns}

\item{log.base}{(numeric) the base of the logarithm}

\item{min.impute}{(numeric) a numeric value to impute the min expression values}
}
\value{
The nomalized expression matrix
}
\description{
log transformation
}
\note{
This function is from SINCERA version a10242015. It will be upgraded soon.
}
