% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{normalization.sqrt.old}
\alias{normalization.sqrt.old}
\title{square root transformation}
\usage{
normalization.sqrt.old(ES)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}
}
\value{
an ExpressionSet object with square root transformed expression values
}
\description{
square root transformation
}
