% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{normalization.sqrt}
\alias{normalization.sqrt}
\title{square root transformation}
\usage{
normalization.sqrt(x)
}
\arguments{
\item{x}{The expression matrix; rows are genes, cells are columns}
}
\value{
The normalized expression matrix
}
\description{
square root transformation
}
\note{
This function is from SINCERA version a10242015. It will be upgraded soon.
}
