% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{normalization.trimmed.mean.old}
\alias{normalization.trimmed.mean.old}
\title{log transformation}
\usage{
normalization.trimmed.mean.old(ES, up = 0.9, lo = 0.1, do.scaling = FALSE)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{up}{(numeric) upper quantile to be removed for calculation of the trimmed mean}

\item{lo}{(numeric) lower quantile to be removed for calculation of the trimmed mean}

\item{do.scaling}{(logical) scale the trimmed mean normalized expression values}
}
\value{
an ExpressionSet object with trimmed mean normalized expression values
}
\description{
log transformation
}
