% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{normalization.trimmed.mean}
\alias{normalization.trimmed.mean}
\title{Trimmed mean normalization}
\usage{
normalization.trimmed.mean(x, up = 0.9, lo = 0.1, do.scaling = FALSE)
}
\arguments{
\item{x}{The expression matrix; rows are genes, cells are columns}

\item{up}{(numeric) upper quantile to be removed for calculation of the trimmed mean}

\item{lo}{(numeric) lower quantile to be removed for calculation of the trimmed mean}

\item{do.scaling}{(logical) scale the trimmed mean normalized expression values}
}
\value{
The normalized expression matrix
}
\description{
Trimmed mean normalization
}
\note{
This function is from SINCERA version a10242015. It will be upgraded soon.
}
