% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{normalization.zscore.1}
\alias{normalization.zscore.1}
\title{Per group zscore normalization of gene expression profiles}
\usage{
normalization.zscore.1(ES, group.by = "GROUP", groups = NULL,
  verbose = TRUE)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the sample information}

\item{groups}{(character) samples for measuring abundancy}

\item{verbose}{(logical) verbose the function output}
}
\value{
an ExpressionSet object with per-sample zscore normalized expression values
}
\description{
Per group zscore normalization of gene expression profiles
}
\note{
This function is from SINCERA version a10242015. It will be upgraded soon.
}
