% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{normalization.zscore.old}
\alias{normalization.zscore.old}
\title{Per group zscore normalization of gene expression profiles}
\usage{
normalization.zscore.old(ES, group.by = SAMPLE.LABEL, groups = NULL,
  verbose = TRUE)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the sample information}

\item{groups}{(character) samples for measuring abundancy}

\item{verbose}{(logical) verbose the function output}
}
\value{
an ExpressionSet object with per-sample zscore normalized expression values
}
\description{
Per group zscore normalization of gene expression profiles
}
