% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{normalization.zscore}
\alias{normalization.zscore}
\title{Performing zscore scaling of gene expression}
\usage{
normalization.zscore(object, pergroup = TRUE, ...)
}
\arguments{
\item{object}{A sincera object}

\item{pergroup}{If true, performing per cell group scaling; otherwise, take all cells as from the same sample and perform scaling.}
}
\value{
An updated sincera object with scaled expression values, use getExpresssion function with scaled=TRUE to access the scaled expression
}
\description{
Perform global or per-sample zscore scaling of gene expression for cell cluster identification and expression pattern visualization
}
