% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{plot.exprs.specificity}
\alias{plot.exprs.specificity}
\title{Plotting the distribution of specificity for each group}
\usage{
plot.exprs.specificity(ES, group.by = SAMPLE.LABEL, groups = NULL,
  ref.idx = NULL, criterion = 0.7,
  specificity.prefix = EXPR.SPECIFICITY.PREFIX, fig.filename = NULL,
  fig.res = 300)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the sample information}

\item{groups}{(character) samples for plotting the distrbution of specificity}

\item{ref.idx}{(numeric) a vectorcontaining the row indices of the set of reference genes, whose specificities will be plotting using green vertical lines}

\item{criterion}{(numeric) a specificity criterion that will be plotted using a red vertical line}

\item{specificity.prefix}{(character) the prefix labeling the columns encoding the results of expression specificity calculation for each group}

\item{fig.filename}{(character) the filename of the output figure; if the fig.filename is null, the figure will be plotted to the screen}

\item{fig.res}{(numeric) resolution of the figure}
}
\description{
Plotting the distribution of specificity for each group
}
