% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCellTypeEnrichment}
\alias{plotCellTypeEnrichment}
\title{Barplot of the cell type enrichment results}
\usage{
plotCellTypeEnrichment(object, groups = NULL, top.k = 5, ...)
}
\arguments{
\item{object}{A sincera object}

\item{groups}{The groups whose cell type enrichment results will be plotted}

\item{top.k}{The number of most enriched cell types to be plotted for each cell group}
}
\description{
Barplot of the cell type enrichment results
}
