% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHC}
\alias{plotHC}
\title{Plot the dendrogram tree of hierarchical clustering}
\usage{
plotHC(object, horiz = FALSE, show.labels = FALSE, do.radial = FALSE, ...)
}
\arguments{
\item{object}{A sincera object with the hc.obj slot stored the dendrogram}

\item{horiz}{If TRUE, plot the dendrogram tree with tips turned right}

\item{show.labels}{If TRUE, show the labels of leafs}

\item{do.radial}{IF TRUE, plot the dendrogram tree using radial layout}
}
\description{
Plot the dendrogram tree of hierarchical clustering
}
