% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap}
\usage{
plotHeatmap(object, genes = NULL, cells = NULL, scaled = T,
  do.log2 = FALSE, order.by.group = T, minmax = c(-1, 1),
  show.labCol = FALSE, show.labRow = FALSE, ...)
}
\arguments{
\item{object}{A sincera object}

\item{genes}{The expression patterns of genes to be shown in the heatmap If NULL, set to all genes}

\item{cells}{The cells to be shown in the heatmap. If NULL, set to all cells}

\item{scaled}{If TRUE, use scaled expression}

\item{do.log2}{IF TRUE, apply log2 transformatoin to the expression}

\item{order.by.group}{IF TRUE, order cells based on their GROUP information}

\item{minmax}{The minimum and maximum values for color scale}

\item{show.labRow}{If TRUE, show the row labels}

\item{show.labcol}{If TRUE, show the column lables}
}
\description{
Plot heatmap
Acknowledgment: Implementation was based on the DoHeatmap function in Seurat
}
