% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPCASD}
\alias{plotPCASD}
\title{Plot the standard deviation of principal components}
\usage{
plotPCASD(object, num.pcs = 20, ...)
}
\arguments{
\item{object}{A sincera object}

\item{num.pcs}{The number of principal components to be plotted}
}
\description{
Plot the standard deviation of principal components
}
