% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotRDS}
\alias{plotRDS}
\title{2D plot of cells in the PCA or tSNE reduced dimensional spaces}
\usage{
plotRDS(object, feature.type = "pca", dim1 = 1, dim2 = 2,
  color.by = "GROUP", pt.size = 5, ...)
}
\arguments{
\item{object}{A sincera object}

\item{feature.type}{The type of components, possible values include pca, tsne}

\item{dim1}{The component for x axis}

\item{dim2}{The component for y axis}

\item{color.by}{The cell metadata used to color cells}

\item{pt.size}{The size of points (cells) in the plot}
}
\description{
2D plot of cells using the PCA or tSNE components
}
