% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{prefiltering}
\alias{prefiltering}
\title{Gene prefiltering based on expression specificity and abundancy}
\usage{
prefiltering(ES, group.by = SAMPLE.LABEL, groups = NULL,
  specificity.threshold = 0.7, abundancy.threshold = 5,
  abundancy.count = 2, mode = "union", do.batch.analysis = FALSE,
  specificity.prefix = EXPR.SPECIFICITY.PREFIX,
  abundancy.prefix = EXPR.ABUNDANCY.PREFIX,
  prefiltering.prefix = PREFILTERING.PREFIX, fig.res = 150, verbose = T,
  export = T, export.components = "fd")
}
\description{
Gene prefiltering based on expression specificity and abundancy
}
