% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{setCellMeta}
\alias{setCellMeta}
\title{Add or update meta data of cells in a sincera object}
\usage{
setCellMeta(object, name, value, verbose = T)
}
\arguments{
\item{object}{A sincera object}

\item{name}{The name of the meta data}

\item{value}{The vector containing the values of the meta data of each cell (should be of the same length as the number of cells in the object)}
}
\value{
The sincera object with the updated cell meta data
}
\description{
Add or update meta data of cells in a sincera object
}
