% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{setCellNames}
\alias{setCellNames}
\title{set the names of all cells in a sincera object}
\usage{
setCellNames(object, value)
}
\arguments{
\item{object}{A sincera object}

\item{value}{The names of all cells}
}
\value{
An updated sincera object
}
\description{
set the names of all cells in a sincera object
}
