% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{setCellType}
\alias{setCellType}
\title{Map cell clusters to cell types}
\usage{
setCellType(object, do.reset = FALSE, groups = NULL, types = NULL)
}
\arguments{
\item{object}{A sincera object}

\item{do.reset}{If TRUE, reset the cell types of all cells to "undefined"}

\item{groups}{A set of cell groups}

\item{types}{The cell types of cell groups}
}
\value{
A sincera object with the updated cell type and cell group mapping
}
\description{
Map cell clusters to cell types
}
