% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{setCellTypeMarkers}
\alias{setCellTypeMarkers}
\title{Add cell type marker information into sincera}
\usage{
setCellTypeMarkers(object, value)
}
\arguments{
\item{object}{A sincera object}

\item{value}{A data frame containing the markers and their corresponding cell types. The first column is cell type, and the second column is marker gene name.}

\item{types}{}
}
\description{
Add cell type marker information into sincera
}
