% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{setExpression}
\alias{setExpression}
\title{Update expression matrix in a sincera object}
\usage{
setExpression(object, value, scaled = FALSE)
}
\arguments{
\item{object}{A sincera object}

\item{value}{An expression matrix, rows are genes, columns are cells, rownames are gene names, colnames are cell names}

\item{scaled}{If TRUE, the update will be performed on the scaled expression matrix, object@sexprs; otherwise, it will be performed on exprs(object@data)}
}
\description{
update expression values of a set of genes in a set of cells in a sincera object
}
