% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{setGeneNames}
\alias{setGeneNames}
\title{Update the gene names in a sincera object}
\usage{
setGeneNames(object, value)
}
\arguments{
\item{object}{A sincera object}

\item{value}{A character vector containing the new names of all genes}
}
\value{
The updated sincera object
}
\description{
Update the gene names in a sincera object
}
