% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{setGenesForClustering}
\alias{setGenesForClustering}
\title{Update the set of genes used for cell cluster identification}
\usage{
setGenesForClustering(object, value)
}
\arguments{
\item{object}{A sincera object}

\item{value}{A vector containing the names of genes for cell cluster identification}
}
\description{
Update the value of the genes.forclustering slot
}
