% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\name{setPCA}
\alias{setPCA}
\title{Update the PCA results in sincera}
\usage{
setPCA(object, name, value)
}
\arguments{
\item{object}{A sincera object}

\item{name}{The name of item to be updated, possible values include rds - principal components, loadings, sdev}

\item{value}{The new value of the item.}
}
\value{
The updated sincera object
}
\description{
update the pca slot
}
