% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{signature.analysis}
\alias{signature.analysis}
\title{Cell type specific signature prediction}
\usage{
signature.analysis(ES, group.by = CLUSTER.LABEL, groups = c(1, 2, 3, 5, 7,
  8, 9), trainset.prefix = TRAINSET.PREFIX,
  train.class.prefix = TRAIN.CLASS.PREFIX, marker.prefix = MARKER.PREFIX,
  testset.prefix = TESTSET.PREFIX, common.prefix = COMMON.PREFIX,
  common.threshold = COMMON.TRESHOLD, common.percentage = COMMON.PERCENTAGE,
  unique.prefix = UNIQUE.PREFIX, unique.ratio = UNIQUE.RATIO,
  unique.quantile = UNIQUE.QUANTILE,
  test.statistic.metric.prefix = TEST.STATS.METRIC.PREFIX, log.base = 2,
  diff.expr.prefix = DIFF.EXPR.PREFIX, diff.expr.threshold = 0.05,
  syn.sim.prefix = SYN.SIM.PREFIX, signature.prefix = SIGNATURE.PREFIX,
  gene.symbol.label = GENE.SYMBOL.LABEL, verbose = TRUE, export = TRUE,
  export.components = "fd", dir.prefix = NULL)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the cluster information}

\item{groups}{(character) the clusters for signature prediction}

\item{trainset.prefix}{(character) prefix for labeling the columns encoding the training instances for each group}

\item{train.class.prefix}{(character) prefix for labeling the columns encoding the class of group specific training instances}

\item{marker.prefix}{(character) prefix for labeling the columns encoding the biomarkers for each group}

\item{testset.prefix}{(character) prefix for labeling the columns encoding the testing instances for each group}

\item{common.prefix}{(character) prefix for labeling the columns encoding the results of common metric for each group}

\item{common.threshold}{(numeric) in common gene metric, expression > COMMON.TRESHOLD will considered as expressed}

\item{common.percentage}{(numeric) in common gene metric, genes that express in >COMMON.PRECENTAGE cluster cells will be considered as a common}

\item{unique.prefix}{(character) prefix for labeling the columns encoding the results of unique metric for each group}

\item{unique.ratio}{(numeric) parameter for unique gene metric}

\item{unique.quantile}{(numeric) parameters for unique gene metric}

\item{test.statistic.metric.prefix}{(character) prefix for labeling the columns encoding the results of test statistic metric for each group}

\item{diff.expr.prefix}{(character) prefix for labeling the columns encoding the results of differential test}

\item{diff.expr.threshold}{(numeric) genes with differential expression p-value<0.05 will be selected as candidates for signature}

\item{syn.sim.prefix}{(character) prefix for labeling the columns encoding the results of synthetic profile similarity metric for each group}

\item{signature.prefix}{(character) prefix for labeling the columns encoding the results of signature prediction}

\item{gene.symbol.label}{(character) the name of the column encoding gene symbols}

\item{verbose}{(logical)}

\item{export}{(logical) wheterh to export the ExpressionSet}

\item{export.components}{(character) the components of an ExpressionSet object that will be exported}
}
\value{
an ExpressionSet object containing the results of signature prediction in the attributes of fData
}
\description{
Cell type specific signature prediction
}
