% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{signature.prediction.old}
\alias{signature.prediction.old}
\title{logistic regression based signature prediction and cross cluster validation
Cell type specific signature prediction}
\usage{
signature.prediction.old(ES, group.by = "CLUSTER", groups = NULL,
  trainset.prefix = TRAINSET.PREFIX,
  train.class.prefix = TRAIN.CLASS.PREFIX, testset.prefix = TESTSET.PREFIX,
  test.statistic.metric.prefix = TEST.STATS.METRIC.PREFIX,
  common.prefix = COMMON.PREFIX, unique.prefix = UNIQUE.PREFIX,
  syn.sim.prefix = SYN.SIM.PREFIX, signature.prefix = SIGNATURE.PREFIX,
  gene.symbol.label = GENE.SYMBOL.LABEL, verbose = TRUE, export = TRUE,
  dir.prefix = "")
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the cluster information}

\item{groups}{(character) the clusters for signature prediction}

\item{trainset.prefix}{(character) prefix for labeling the columns encoding the training instances for each group}

\item{train.class.prefix}{(character) prefix for labeling the columns encoding the class of group specific training instances}

\item{testset.prefix}{(character) prefix for labeling the columns encoding the testing instances for each group}

\item{test.statistic.metric.prefix}{(character) prefix for labeling the columns encoding the results of test statistic metric for each group}

\item{common.prefix}{(character) prefix for labeling the columns encoding the results of common metric for each group}

\item{unique.prefix}{(character) prefix for labeling the columns encoding the results of unique metric for each group}

\item{syn.sim.prefix}{(character) prefix for labeling the columns encoding the results of synthetic profile similarity metric for each group}

\item{signature.prefix}{(character) prefix for labeling the columns encoding the results of signature prediction}

\item{gene.symbol.label}{(character) the name of the column encoding gene symbols}

\item{verbose}{(logical)}

\item{export}{(logical) wheterh to export the ExpressionSet}

\item{export.components}{(character) the components of an ExpressionSet object that will be exported}
}
\value{
an ExpressionSet object containing the results of signature prediction in the attributes of fData
}
\description{
logistic regression based signature prediction and cross cluster validation
Cell type specific signature prediction
}
