% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{signature.prediction}
\alias{signature.prediction}
\title{Predicting cell type signature}
\usage{
signature.prediction(object, groups = NULL, use.logireg = FALSE,
  sigs.n = NULL, diff.method = "welch", use.fdr = FALSE,
  diff.thresh = 0.05, avg.thresh = 1, min.expression = 1,
  pct.thresh = 0.8, fc.thresh = 1.5, ...)
}
\arguments{
\item{object}{A sincera object}

\item{groups}{The cell groups for signature prediction}

\item{use.logireg}{If TRUE, use logistic regression model based method to predict signature for cell groups (cell types) with marker information available;
use setCellTypeMarkers to add marker information to sincera, and use getCellTypeMarkers to check added cell type information}

\item{sigs.n}{A vector containing the number of signatures to be predicted for each cell group; if NULL, use genes with top 20 prediction score as the signature}

\item{diff.method}{Take effect when use.logireg==FALSE. The method of differentiatial expression}

\item{use.fdr}{Take effect when use.logireg==FALSE. If TRUE, use BH to adjust pvalues of differential expression}

\item{diff.thresh}{Take effect when use.logireg==FALSE. The threshold of pvalue or fdr.}

\item{avg.thresh}{A cell type signature gene must have at least avg.thresh expression in the defined cell type}

\item{min.expression}{The threshold for determining expressed genes}

\item{pct.thresh}{A cell type signature gene must express in at least pct.thresh percentage of the cells of the defined cell type}

\item{fc.thresh}{A cell type signature gene must have at least fc.thresh fold of average expression induction in its defined cell type cells when compared to its average expression in all the other cells}
}
\value{
An updated sincera object, use getSigGenes to access the predicted signature genes'
}
\description{
By default, the function predicts cell type signature by using percentage filter and fold-induction filter to refine cluster specific differentially expressed genes
When the marker information for some cell types is available, one can choose to predict signature using a logistic regression model integrating four metrics
}
