% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{signature.validation.old}
\alias{signature.validation.old}
\title{Repeated Random Subsampling based Validation of Signature Prediction}
\usage{
signature.validation.old(ES, group.by = "CLUSTER", groups = NULL,
  diff.threshold = 0.05, sig.n = NULL, percentages = c(0.2),
  repeats = 10, diff.expr.prefix = DIFF.EXPR.PREFIX,
  signature.prefix = SIGNATURE.PREFIX, verbose = T, export = T)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the cluster information}

\item{groups}{(character) the clusters for signature prediction}
}
\description{
Repeated Random Subsampling based Validation of Signature Prediction
}
