% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{signature.validation}
\alias{signature.validation}
\title{Validating cell type signature using random subsampling mehod}
\usage{
signature.validation(object, siggenes = NULL, pct = 0.2, repeats = 3, ...)
}
\arguments{
\item{object}{A sincera object}

\item{siggenes}{A data frame containing the . The first column of the data frame contains cell cluster information, while the second column contains the genes}

\item{pct}{The percentage of cells to be used to construct testing set}

\item{repeats}{The number of subsamplings}
}
\value{
An updated sincera object, use getSigValidation to access the validation results
}
\description{
Validating cell type signature using random subsampling mehod
}
