% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{sigpred.testing.sets}
\alias{sigpred.testing.sets}
\title{Constructing group specific testing sets for signature prediction}
\usage{
sigpred.testing.sets(ES, group.by = CLUSTER.LABEL, groups = NULL,
  testset.prefix = TESTSET.PREFIX, diff.expr.prefix = DIFF.EXPR.PREFIX,
  diff.expr.threshold = 0.05)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the cluster information}

\item{groups}{(character) the clusters for signature prediction}

\item{testset.prefix}{(character) prefix for labeling the columns encoding the testing instances for each group}

\item{diff.expr.prefix}{(character) prefix for labeling the columns encoding the results of differential test}

\item{diff.expr.threshold}{(numeric) genes with differential expression p-value<0.05 will be selected as candidates for signature}
}
\value{
an ExpressionSet object containing the constructed testing sets in the attributes of fData
}
\description{
Constructing group specific testing sets for signature prediction
}
