% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{sigpred.training.sets}
\alias{sigpred.training.sets}
\title{Constructing group specific training sets for signature prediction}
\usage{
sigpred.training.sets(ES, group.by = CLUSTER.LABEL, groups = NULL,
  trainset.prefix = TRAINSET.PREFIX,
  train.class.prefix = TRAIN.CLASS.PREFIX,
  test.statistic.metric.prefix = TEST.STATS.METRIC.PREFIX,
  common.prefix = COMMON.PREFIX, unique.prefix = UNIQUE.PREFIX,
  syn.sim.prefix = SYN.SIM.PREFIX, marker.prefix = MARKER.PREFIX)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the cluster information}

\item{groups}{(character) the clusters for signature prediction}

\item{trainset.prefix}{(character) prefix for labeling the columns encoding the training instances for each group}

\item{train.class.prefix}{(character) prefix for labeling the columns encoding the class of group specific training instances}

\item{test.statistic.metric.prefix}{(character) prefix for labeling the columns encoding the results of test statistic metric for each group}

\item{common.prefix}{(character) prefix for labeling the columns encoding the results of common metric for each group}

\item{unique.prefix}{(character) prefix for labeling the columns encoding the results of unique metric for each group}

\item{syn.sim.prefix}{(character) prefix for labeling the columns encoding the results of synthetic profile similarity metric for each group}

\item{marker.prefix}{(character) prefix for labeling the columns encoding the biomarkers for each group}
}
\value{
an ExpressionSet object containing the constructed training sets in the attributes of fData
}
\description{
Constructing group specific training sets for signature prediction
}
