% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincera.R
\docType{class}
\name{sincera}
\alias{sincera}
\alias{sincera-class}
\title{The sincera Class}
\description{
The object storing all information associated with the sincera analysis, including data, annotations, analyes, etc.
It is recommended to use the defined get and set methods to access and manipulate the data in the sincera object.
}
\details{
Information is stored in slots. Key slots include:
}
\section{Slots}{

 \describe{
   \item{\code{data}:}{\code{"ExpressionSet"}, the scRNA-seq data in ExpressionSet format, encoding the expression matrix, gene, and cell metadata, including sample and condition information. In most of the SINCERA functions, cell grouping is based on the "GROUP" metadata.  }
   \item{\code{genes.forclustering}:}{\code{"vector"}, the set of genes used for the cell cluster identification  }
   \item{\code{difftests}:}{\code{"list"}, results of differential expression test of all genes in the data using different methods}
   \item{\code{siggenes}:}{\code{"data.frame"}, a data frame containing the significant differentially expressed genes   }
   \item{\code{cte}:}{\code{"list"}, results of cell type enrichment  }
   \item{\code{ctmap}:}{\code{"data.frame"}, a data frame containing the mapping between cell clusters and cell types  }
   \item{\code{markers}:}{\code{"data.frame"}, a data frame containing known cell type markers }
   \item{\code{siggenes}:}{\code{"data.frame"}, a data frame containing the signature genes based on signature prediction   }
   \item{\code{sigvalidate}:}{\code{"data.frame"}, a data frame containing the results of signature validation   }
   \item{\code{df}:}{\code{"list"}, a list containing the results of driving force analysis   }
   \item{\code{associations}:}{\code{"data.frame"}, a data frame containing the gene and cell type association table   }
   \item{\code{tfs}:}{\code{"vector"}, a vector containing a set of all transcription factors/cofactors   }
   \item{\code{projname}:}{\code{"character"}, a string describing the data and analysis  }

}
}

