% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{specificity.criterion.selection}
\alias{specificity.criterion.selection}
\title{Determine the threshold for specificity filter based on a set of reference genes}
\usage{
specificity.criterion.selection(ES, group.by = SAMPLE.LABEL, groups = NULL,
  ref.idx = NULL, exp.t = 5, exp.p = 0.95, tolerance = 0.05,
  step = 0.1, do.plot = FALSE, verbose = T)
}
\arguments{
\item{ES}{(ExpressionSet) an ExpressionSet object containing the single cell RNA-seq data}

\item{group.by}{(character) the name of the column that contains the sample information}

\item{groups}{(character) samples involves in determining the threshold}

\item{ref.idx}{(numeric) a vectorcontaining the row indices of the set of reference genes}

\item{exp.t}{(numeric) min expression value for the reference genes}

\item{exp.p}{(numeric) min expression percentage per sample for the reference genes}

\item{tolerance}{(numeric) percentage of reference genes that will be allowed to pass the specificity filters}

\item{step}{(numeric) a numeric value between 0 and 1, specifying the granularity of}

\item{do.plot}{(logical) plotting the distribution of specificity, the specificities of the reference genes (green lines), and the selected threshold (red line, if exists)}

\item{verbose}{(logical) verbose the function output}
}
\value{
a numeric value between 0 and 1 if a threshold exists, otherwise, -1
}
\description{
Determine the threshold for specificity filter based on a set of reference genes
}
