% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{specificity.thresholdSelection}
\alias{specificity.thresholdSelection}
\title{Selecting a threshold value for specificity}
\usage{
specificity.thresholdSelection(object, genes)
}
\arguments{
\item{object}{A sincera object}

\item{genes}{The set of housekeeping genes used to determine the specificity threshold}
}
\description{
Selecting a threshold value for specificity
}
