% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{celltype.enrichment}
\alias{celltype.enrichment}
\title{Perform cell type enrichment}
\usage{
celltype.enrichment(object, species = "MUSMU", id.type = "SYMBOL", ...)
}
\arguments{
\item{object}{A sincera oject}

\item{species}{The species of the data, possible values include MUSMU - mouse, HOMSA - human}

\item{id.type}{The type of gene identifier, possible values include SYMBOL - Entrez SYMBOL, EG - Entrez ID, ENSEMBL - Ensembl ID}
}
\value{
An updated sincera object with enrichment results in the cte slot
}
\description{
Perform cell type enrichment
}
