% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{cluster.assignment}
\alias{cluster.assignment}
\title{Finding cell clusters}
\usage{
cluster.assignment(object, feature.type = "gene", rds.dims = 1:3,
  update.cellgroup = T, clustering.method = "hc", verbose = T, ...)
}
\arguments{
\item{object}{A SINCERA object}

\item{feature.type}{The feature space used for clustering: "gene" - means gene expression space, "pca" - means reduced dimension space using PCA}

\item{rds.dims}{A numberic vector specified the reduced dimensions used for clustering. Only take effect when the feature.type is "pca".}

\item{update.cellgroup}{The clustering result will be saved to the "CLUSTER" meta data. If TRUE, the GROUP meta data will be updated as well.}

\item{clustering.method}{The cluster identification algorithm, possible values include pam, kmeans, hc - hierarchical clustering, graph - kNN + Community detection, tight, consensus}

\item{verbose}{If TRUE, print the verbose messages}
}
\value{
The update sincera object with clustering results in the "CLUSTER" meta data, use getCellMeata with name="CLUSTER" to assess the result
}
\description{
Finding cell clusters
}
