% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMarkers}
\alias{plotMarkers}
\title{Plot gene expression patterns}
\usage{
plotMarkers(object, genes, use.scaled = T, do.log2 = FALSE, do.order = T,
  show.jitter = T, plots = c("cell", "boxplot", "pca", "tsne"),
  dims.use = 1:2, pt.size = 1, font.size = 8, ...)
}
\arguments{
\item{object}{A sincera object}

\item{genes}{The expression patterns of genes to be plotted}

\item{use.scaled}{If TRUE, use the scaled expression values for plotting}

\item{do.log2}{If TRUE, apply log2 transformation to the values}

\item{do.order}{If TRUE, order cells based on their GROUP information}

\item{show.jitter}{If TRUE, in the violin and boxplot, plot individual cells}

\item{plots}{The type of plots to be shown, possible values include cell, violin, boxplot, pca, and tsne}

\item{dims.use}{A numeric vector containing the indices of the two dimensions for pca or tsne plot}

\item{pt.size}{The size of points in the plot}

\item{fontsize}{The size of font in the plot}
}
\description{
Plot gene expression patterns
}
